+++ glsl/Experimental-Effects/Anime4K_ThinLines_HQ.glsl
@@ -112,15 +112,15 @@ vec4 hook() {
 #define SIGMA (HOOKED_size.y / 1080.0) * 2.0
 #define KERNELSIZE (SIGMA * 2.0 + 1.0)
 
-float gaussian(int x, float s, int m) {
-	return (1.0 / (s * sqrt(2.0 * 3.14159))) * exp(-0.5 * pow(abs(float(x - m)) / s, 2.0));
+float gaussian(float x, float s, float m) {
+	return (1.0 / (s * sqrt(2.0 * 3.14159))) * exp(-0.5 * pow(abs(x - m) / s, 2.0));
 }
 
 float lumGaussian(vec2 pos, vec2 d) {
-	float g = (L_tex(pos).x) * gaussian(0, SIGMA, 0);
-	g = g + (L_tex(pos - d).x + L_tex(pos + d).x) * gaussian(1, SIGMA, 0);
+	float g = (L_tex(pos).x) * gaussian(0.0, SIGMA, 0.0);
+	g = g + (L_tex(pos - d).x + L_tex(pos + d).x) * gaussian(1.0, SIGMA, 0.0);
 	for (int i=2; float(i)<KERNELSIZE; i++) {
-		g = g + (L_tex(pos - (d * float(i))).x + L_tex(pos + (d * float(i))).x) * gaussian(i, SIGMA, 0);
+		g = g + (L_tex(pos - (d * float(i))).x + L_tex(pos + (d * float(i))).x) * gaussian(float(i), SIGMA, 0.0);
 	}
 	
 	return g;
@@ -144,15 +144,15 @@ vec4 hook() {
 #define SIGMA (HOOKED_size.y / 1080.0) * 2.0
 #define KERNELSIZE (SIGMA * 2.0 + 1.0)
 
-float gaussian(int x, float s, int m) {
-	return (1.0 / (s * sqrt(2.0 * 3.14159))) * exp(-0.5 * pow(abs(float(x - m)) / s, 2.0));
+float gaussian(float x, float s, float m) {
+	return (1.0 / (s * sqrt(2.0 * 3.14159))) * exp(-0.5 * pow(abs(x - m) / s, 2.0));
 }
 
 float lumGaussian(vec2 pos, vec2 d) {
-	float g = (L_tex(pos).x) * gaussian(0, SIGMA, 0);
-	g = g + (L_tex(pos - d).x + L_tex(pos + d).x) * gaussian(1, SIGMA, 0);
+	float g = (L_tex(pos).x) * gaussian(0.0, SIGMA, 0.0);
+	g = g + (L_tex(pos - d).x + L_tex(pos + d).x) * gaussian(1.0, SIGMA, 0.0);
 	for (int i=2; float(i)<KERNELSIZE; i++) {
-		g = g + (L_tex(pos - (d * float(i))).x + L_tex(pos + (d * float(i))).x) * gaussian(i, SIGMA, 0);
+		g = g + (L_tex(pos - (d * float(i))).x + L_tex(pos + (d * float(i))).x) * gaussian(float(i), SIGMA, 0.0);
 	}
 	
 	return g;
